Add typealias support examples
This required slight modification to the code generation in order to properly
use the underlying type of an alias in certain locations and the name of
the typealias in others. Additionally, had to remove support for unwrapping
iterables from the string wrapper class but there aren't many
use cases for that feature that I can think of.